+++ src/school/__init__.py
@@ -5,6 +5,6 @@ __all__ = ["Classroom", "Result", "Task", "Plotter", "import_results", "export_r
 
 from .task import Task
 from .result import Result
-from .io import *
+from .imp_exp import *
 from .plotter import Plotter
 from .classroom import Classroom
